Fixed command-line toolset configuration.
The toolsets must be defined in project-config.jam in order to allow the
toolset to be set on the command line.  This change is less than ideal
since the toolsets to test must be specified in two places, but it is
worth doing because limiting the toolsets on the command line is
extremely important.